moved typeDefs and resolvers to schema directory
