+++ server/server.ts
@@ -1,8 +1,8 @@
 import { ApolloServer } from "apollo-server-express";
 
 import ArtistAPI from "./data/ArtistAPI";
-import { typeDefs } from "./typeDefs";
-import { resolvers } from "./resolvers";
+import { typeDefs } from "./schema/typeDefs";
+import { resolvers } from "./schema/resolvers";
 
 export const server = new ApolloServer({
   typeDefs,
